% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfReg.R
\name{dfReg}
\alias{dfReg}
\title{This is the main function in the dfReg package,}
\usage{
dfReg(regression_model, guess_model_type = NULL, ...)
}
\arguments{
\item{regression_model}{The inputed model to be processed}

\item{guess_object_type}{If type not supported can try to guess from a similar object. Currently knows c("lm", "glm", "merModLmerTest", "lmerMod")}
}
\value{
a data.frame object with predictor names and estimates
}
\description{
This function takes in a fitted model objects, determines the type of the object, and processes it accordingly
}
\examples{
test_data = data.frame(predictor = 1:10, outcome = rnorm(10))
test_model = lm(outcome ~ predictor, data = test_data)
dfReg(test_model)
}

