% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_df.R
\name{extract_df}
\alias{extract_df}
\title{This function directs to the correct processing function}
\usage{
extract_df(regression_model, object_type, ...)
}
\arguments{
\item{regression_model}{The inputed model to be processed}

\item{object_type}{the string argument to be evaluated}
}
\value{
a data.frame object with predictor names and estimates
}
\description{
This function directs to the correct processing function
}

