% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract_df_regressiontype.R
\name{extract_df_lm}
\alias{extract_df_lm}
\title{This function processed lm or glm regression}
\usage{
extract_df_lm(lm_model, include.intercept = T, include.confint = F)
}
\arguments{
\item{lm_model}{The inputed lm or glm model to be processed}

\item{include.intercept}{Include the intercept estimate. Default TRUE.}

\item{include.confint}{Include the confidence interval estimates. Default FALSE.}
}
\value{
a data.frame object with predictor names and estimates
}
\description{
This function processed lm or glm regression
}
\examples{
test_data = data.frame(predictor = 1:10, outcome = rnorm(10))
test_model = lm(outcome ~ predictor, data = test_data)
extract_df_lm(test_model)
extract_df_lm(test_model, include.intercept = F)
}

