% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_df_regressiontype.R
\name{extract_df_lmerTest}
\alias{extract_df_lmerTest}
\title{This function processed lmer model when lmerTest package is loaded}
\usage{
extract_df_lmerTest(lmerTest_model, include.intercept = T)
}
\arguments{
\item{lmerTest_model}{A model generated using lmerTest}

\item{include.intercept}{Include the intercept estimate. Default TRUE.}
}
\value{
a data.frame object with predictor names and estimates
}
\description{
This function processed lmer model when lmerTest package is loaded
}
\examples{
library(lmerTest)
test_data = data.frame(
  predictor1 = 1:100,
  predictor2 = rep(LETTERS[1:2], each = 50),
  outcome = rnorm(100),
  group = sample(LETTERS[1:3], 100, replace = T)
  )
test_model = lmer(outcome ~ predictor1 + predictor2 + (1|group), data = test_data)
extract_df_lmerTest(test_model)
extract_df_lmerTest(test_model, include.intercept = F)

}

