% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_object_type.R
\name{get_object_type}
\alias{get_object_type}
\title{This function takes a model object and return its type}
\usage{
get_object_type(regression_model, position = 1)
}
\arguments{
\item{regression_model}{The inputted model to be processed}
}
\value{
a string of type
}
\description{
This function takes a model object and return its type
}

