% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_type_call.R
\name{object_type_call}
\alias{object_type_call}
\title{This is the main function in the dfReg package,}
\usage{
object_type_call(regression_model, position = 1)
}
\arguments{
\item{regression_model}{The inputted model to be processed}
}
\value{
a string of type
}
\description{
This function takes a model object and return its type
}

