% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract_df.R
\name{extract_df}
\alias{extract_df}
\title{This is the main function in the dfReg package,}
\usage{
extract_df(regression_model, guess_model_type = NULL, include.intercept = T,
  include.confint = F, ...)
}
\arguments{
\item{regression_model}{The inputed model to be processed}

\item{include.confint}{Include the confidence interval estimates. Default FALSE.}

\item{guess_object_type}{If type not supported can try to guess from a similar object. Currently knows c("lm", "glm", "merModLmerTest", "lmerMod")}
}
\value{
a data.frame object with predictor names and estimates
}
\description{
This function takes in a fitted model objects, determines the type of the object, and processes it accordingly
}
\examples{
test_data = data.frame(predictor = 1:10, outcome = rnorm(10))
test_model = lm(outcome ~ predictor, data = test_data)
extract_df(test_model)
}

