% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_df_regressiontype.R
\name{extract_df_lm}
\alias{extract_df_lm}
\title{This function processed lm or glm regression}
\usage{
extract_df_lm(regression_model, include.intercept = T)
}
\arguments{
\item{regression_model}{The inputed model to be processed}

\item{include.intercept}{Include the intercept estimate. Default TRUE.}
}
\value{
a data.frame object with predictor names and estimates
}
\description{
This function processed lm or glm regression
}
\examples{
test_data = data.frame(predictor = 1:10, outcome = rnorm(10))
test_model = lm(outcome ~ predictor, data = test_data)
extract_df_lm(test_model)
extract_df_lm(test_model, include.intercept = F)

}

